refactor!: code changes and optimisations
