+++ requirements.txt
@@ -1,6 +1,6 @@
 thefuzz
 fastapi[all]
-devgoldyutils
+devgoldyutils>=2.5.8
 typing-extensions
 uvicorn[standard]
 slowapi
